#205 - Restructure nav bar files
