+++ src/components/shared/nav-top-bar/nav-page-links/nav-page-links.test.tsx
@@ -4,7 +4,7 @@
  */
 
 import { screen } from '@testing-library/react';
-import { renderWithRouter } from '../../../test-support/test-utils';
+import { renderWithRouter } from '../../../../test-support/test-utils';
 import NavPageLinks from './nav-page-links';
 
 /**
